requirements.txt
onnxruntime
openvino
flask
numpy
pillow
watchdog==6.0.0